+++ lib/cisco_node_utils/encapsulation.rb
@@ -71,7 +71,7 @@ module Cisco
       config_set('encapsulation', 'create', profile: @encap_name)
     end
 
-    # Destroy a encap instance; disable feature on last instance
+    # Destroy an encap instance
     def destroy
       config_set('encapsulation', 'destroy', profile: @encap_name)
     end
@@ -81,22 +81,28 @@ module Cisco
     # ----------
 
     def range_summarize(string)
-      Utils.array_to_string(Encapsulation.string_to_array(string.to_s), false)
+      Utils.array_to_str(Encapsulation.string_to_array(string.to_s), false)
     end
 
     def dot1q_map
       result = config_get('encapsulation', 'dot1q_map', profile: @encap_name)
-      return default_dot1q_map if result.to_s.empty?
+      return default_dot1q_map if result.empty?
+
       result[0] = range_summarize(result[0])
       result[1] = range_summarize(result[1])
       result
     end
 
-    def dot1q_map=(val)
-      no_cmd = (val.empty?) ? 'no' : ''
-      val = dot1q_map if val.empty?
+    def dot1q_map=(map)
+      state = ''
+      if map.empty?
+        state = 'no'
+        map = dot1q_map
+        return if map.empty?
+      end
+      vlans, vnis = map
       config_set('encapsulation', 'dot1q_map', profile: @encap_name,
-                 state: no_cmd, vlans: val[0], vnis: val[1])
+                 state: state, vlans: vlans, vnis: vnis)
     end
 
     def default_dot1q_map